// packages/cloudpivot/api/src/application.ts
export interface IsMobileSchema {
  isMobile: boolean;
  userId?: string;
  trustType?: string;
  isPortal?: boolean;
}

export interface FolderSchema {
  appCode: string;
  isMobile: boolean;
}

export interface folderIdSchema {
  id: string;
}

export interface ListConfigSchame {
  code: string;
  schemaCode: string;
  source?: number;
  clientType?: string;
}
// ...
export interface QueryHeaderParams {
  schemaCode: string;
  clientType: QueryClientType;
}

export interface QueryListParams {
  mobile: boolean;
  page: number;
  size: number;
  queryCode: string;
  schemaCode: string;
  filters?: Array<Filters>;
  queryCondition?: any[];
  queryVersion?: number;
  orderByFields?: string[];
  orderType?: number;
  options?: QueryListOptions;
  formCode?: string;
  reverseCode?: string;
  reverseSchemaCode?: string;
  reverseSchemaParam?: any;
  reverseViewFlag?: boolean;
  QueryDataVo?: boolean | string;
  tempAuthSheetId?: string;
  tempAuthObjectId?: string;
  tempAuthPropertyCode?: string;
}

export interface QuerydynamicParams {
  mobile: boolean;
  userName: string;
  corpId: string;
  dynamic: string;
  page?: number;
  size?: number;
}

export interface FormUrlParams {
  bizObjectId: string;
  schemaCode: string;
  formCode?: string;
}

export interface Filters {
  propertyCode: string;
  propertyType: number;
  propertyValue: string;
}

export interface ApiImplicitParam {
  schemaCode: string;
  relativeCode?: string;
  isMobile: boolean;
  queryActionType: number;
  queryActionRelativeType?: number;
}

export interface fileNameType {
  fileName: string;
}

export enum QueryClientType {
  /**
   *客户端类型--PC
   */
  PC = 'PC',

  /**
   * 客户端类型--MOBILE
   */
  APP = 'APP',

  /**
   * 客户端类型--OTHER
   */
  OTHER = 'OTHER',
}

export enum QueryDisplayType {
  /**
   * 默认展示查询列表配置的显示字段
   */
  Default = '0',

  /**
   * 覆盖展示查询列表配置的显示字段
   */
  Override = '1',

  /**
   * 在查询列表配置的显示字段的基础上新增显示字段
   */
  Append = '2',
}

export interface QueryListOptions {
  customDisplayColumns: string[];

  queryDisplayType: QueryDisplayType;
}

/**
 * 下载示例文件请求参数
 */
export interface ExportTemplateParams {
  schemaCode: string;
  queryCode: string;
}

/**
 * 导入请求参数
 */
export interface SecondImportParams {
  headColumns: Array<any>;
  secondImportData: Array<any>;
  queryField: string;
}
export interface SheetSecondImportParams {
  headColumns: Array<any>;
  secondImportData: Array<any>;
}

export interface ImportParams {
  fileName: string;
  originalFilename: string;
  schemaCode: string;
  queryCode: string;
  queryField: any;
  languageType: string;
}

/**
 * 导入结果代码
 */
export enum ImportResult {
  None = -1,
  Success = 0, // 导入成功
  PartialSuccess = 1, // 导入部分成功
  FileTypeError = 2, // 文件类型错误
  DataNumExceed = 3, // 数据超过500限制
  DataColumnError = 4, // 文件展示头信息错误
  SystemError = 5, // 系统错误
  Unspecified = 6, // 表单关联了流程实例
  TemplateEmpty = 7, // 模板为空，请填写信息
  FaileReImport = 8, // 选人数据项填入的值不唯一或者为空，请确定后再导入这部分内容
}
export interface DeleteListParams {
  ids: Array<string>;
  schemaCode: string;
  queryId?: string; //甘特图删除时添加
}

export interface savePaperParams {
  list: Array<string>;
}

export interface ExportDataParams {
  filters: Array<Filters>;
  allObjectId?: any;
  columns?: any;
  objectIds?: any;
  mobile: boolean;
  page: number;
  queryCode: string;
  schemaCode: string;
  size: number;
  orderByFields?: string[];
  orderType?: number;
  view?: any;
  showTotal?: boolean;
}

export interface ExportId {
  id: string;
}

export interface ExportHistoryParam {
  type: string;
  schemaCode: string;
  queryCode: string;
}

export interface WorkflowSchemaCode {
  schemaCode: string;
  userId?: string;
  trustType?: string;
  isPublish?: boolean;
  isPortal?: boolean;
}

export interface AppCode {
  appCode: string;
  trustType?: string;
  isPortal?: boolean;
}

export interface SearchBizModelRequest {
  appCode?: string;
  searchKey: string;
}

export interface AppPackageSearch {
  name: string;

  appCode?: string;
}

export interface GenShortCodesParams {
  pairSettingModels: Array<ShortCodeParams>;
}

interface ShortCodeParams {
  pairValue: string;
}

export interface GetReportParams {
  code: string;
}

export interface FormItem {
  schemaCode: string;
  id: string;
  data: FormData;
  instanceId: string;
  sequenceStatus: string;
  modifiedTime: string;
  return?: any;
}

interface FormData {
  instanceId: string;
  startTime: string;
  finishedTime: string;
  imgUrl: string;
  userName: string;
  name: string;
  sheetCode: string;
  modifiedTime: string;
  createdTime: string;
}

export interface appManagerModel {
  // 定义获取子管理员管理的应用模型过滤参数
  filterType?: string;
}

export interface AppCodeByModelCodeParam {
  schemaCode: string;
}